Проверил роуты и контроллеры через Postman

+++ controllers/movies.js
@@ -2,6 +2,8 @@ import mongoose from 'mongoose';
 import Movie from '../models/movie.js';
 import ValidationError from '../errors/validation-error.js';
 import NotFoundError from '../errors/not-found-error.js';
+import AlreadyExistError from '../errors/already-exist-error.js';
+import AccessError from '../errors/access-error.js';
 
 export async function getMovies(req, res, next) {
   try {
@@ -33,6 +35,10 @@ export async function addMovie(req, res, next) {
       next(new ValidationError(err));
       return;
     }
+    if (err.code === 11000) {
+      next(new AlreadyExistError('Этот фильм уже добавлен в избранное'));
+      return;
+    }
 
     next(err);
   }
@@ -41,7 +47,11 @@ export async function addMovie(req, res, next) {
 export async function deleteMovie(req, res, next) {
   try {
     const { movieId } = req.params;
-    await Movie.findById({ _id: movieId }).orFail();
+    const userId = req.user._id;
+    const movie = await Movie.findById({ _id: movieId }).orFail();
+    if (movie.owner._id.toString() !== userId) {
+      throw new AccessError('У вас нет прав для удаления этого фильма');
+    }
     const data = await Movie.deleteOne({ _id: movieId });
     res.status(200).send({ data });
   } catch (err) {
